+++ packages/blocks/src/Navigation/DesktopNavigation.tsx
@@ -1,8 +1,9 @@
 import React, { FC } from 'react';
-import { Flex, Image, Link } from '@cmpsr/components';
+import { Flex, Link } from '@cmpsr/components';
 import { NavigationProps } from './types';
+import { LogoNavigation } from './LogoNavigation';
 
-export const DesktopNavigation: FC<Omit<NavigationProps, 'sticky'>> = ({ actions, anchors, logoUrl }) => (
+export const DesktopNavigation: FC<Omit<NavigationProps, 'sticky'>> = ({ actions, anchors, logoUrl, logoHref }) => (
   <Flex
     data-testid="desktop-navigation"
     maxWidth="80rem"
@@ -12,7 +13,7 @@ export const DesktopNavigation: FC<Omit<NavigationProps, 'sticky'>> = ({ actions
     width="100%"
   >
     <Flex gap={{ lg: '2.75rem', xxl: '4.5rem' }}>
-      <Image src={logoUrl} alt="navigation logotype" height="2rem" />
+      <LogoNavigation logoHref={logoHref} logoUrl={logoUrl} />
       <Flex gap={{ lg: '1.5rem', xl: '2.25rem', xxl: '5rem' }}>
         {anchors.map(({ label, href }, i: number) => (
           <Link href={href} size="m" key={i} color="text-link-secondary-default">