+++ packages/blocks/src/index.ts
@@ -1,4 +1,5 @@
 export * from './BrandLogos';
 export * from './CookieBanner';
+export * from './Hero';
 export * from './Navigation';
 export * from './Testimonial';