+++ packages/components/src/components/primitives/TextPairing/TextPairing.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta } from '@storybook/react';
-import { TextPairing, TextPairingProps, textPairingVariants, textPairingTextAlignments } from '.';
+import { TextPairing, textPairingVariants, textPairingTextAlignments } from '.';
 import { Text } from '@components';
 import { Box, Flex, Spacer, Stack, StackDivider } from '@chakra-ui/layout';
 import { defaultColors } from '@theme/ComposerTheme/Colors';
@@ -28,7 +28,7 @@ export default {
   },
 } as Meta;
 
-const AllTemplate = ({ label, subLabel, textAlign }: TextPairingProps) => (
+const AllTemplate = ({ label, subLabel, textAlign }) => (
   <Stack spacing="2rem" py="2rem" divider={<StackDivider borderColor="#888" />}>
     {textPairingVariants.map((variant) => (
       <Flex key={variant} px="1rem">
@@ -37,7 +37,12 @@ const AllTemplate = ({ label, subLabel, textAlign }: TextPairingProps) => (
         </Box>
         <Spacer />
         <Box>
-          <TextPairing variant={variant} label={label} subLabel={subLabel} textAlign={textAlign} />
+          <TextPairing
+            variant={variant}
+            labelProps={{ children: label }}
+            subLabelProps={{ children: subLabel }}
+            textAlign={textAlign}
+          />
         </Box>
       </Flex>
     ))}
@@ -49,10 +54,11 @@ All.args = {
   subLabel: 'Oh, yeah!',
 };
 
-const ResponsiveTemplate = (args: TextPairingProps) => (
+const ResponsiveTemplate = () => (
   <Box py="2rem">
     <TextPairing
-      {...args}
+      labelProps={{ children: 'Label' }}
+      subLabelProps={{ children: 'SubLabel' }}
       variant={{
         base: 'textpairing-body-bold',
         md: 'textpairing-header-L',
@@ -62,14 +68,15 @@ const ResponsiveTemplate = (args: TextPairingProps) => (
   </Box>
 );
 export const ResponsiveVariant = ResponsiveTemplate.bind({});
-ResponsiveVariant.args = {
-  label: 'Label',
-  subLabel: 'SubLabel',
-};
 
-const Template = (args: TextPairingProps) => (
+const Template = ({ label, subLabel, variant, textAlign, otherLabelProps = {}, otherSubLabelProps = {} }) => (
   <Box py="2rem">
-    <TextPairing {...args} />
+    <TextPairing
+      labelProps={{ children: label, ...otherLabelProps }}
+      subLabelProps={{ children: subLabel, ...otherSubLabelProps }}
+      variant={variant}
+      textAlign={textAlign}
+    />
   </Box>
 );
 export const Playground = Template.bind({});
@@ -77,4 +84,7 @@ Playground.args = {
   variant: textPairingVariants[0],
   label: 'Label',
   subLabel: 'SubLabel',
+  otherLabelProps: { color: 'text-primary' },
+  otherSubLabelProps: { color: 'text-primary' },
+  textAlign: 'start',
 };