+++ packages/components/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cmpsr/components
 
+## 3.1.0
+
+### Minor Changes
+
+- c06accb: feat: add breakpoints and styles props to Theme
+
+## 3.0.0
+
+### Major Changes
+
+- f292607: feat: feat: allow TextPairing to receive all Text props"
+
+## 2.4.0
+
+### Minor Changes
+
+- adba0b1: feat: add subLabel whiteSapce prop to TextPairing"
+
 ## 2.3.0
 
 ### Minor Changes